feat: Page mnges components to refresh effectively
Featuring... cyclic dependencies between Component <--> Page (yuck) and clusteringly updating the fixed parts of each component from each page effectively.

Instead of taking in pageSelector(string) as the param we take in page(Page) so that the component will be automatically added to each page's component stack in constructor.

When scrolling, last pages items will be set to invisible (disappear) and vice versa for the next page