fix: Eliminated mouse buggy bhvr of chevrons
The buggy behaviour shows that if the mouse enters the trigger zone of the chevron as the next page is not fully transitioned, the scroll speed of the component would become chaotic.

A solution would be to only make the chevrons load up once the transition is fully complete. This has already been accomplished in the previous commits, but not when the page is loaded for the FIRST time.