test: State mnger for getter cls to get pg info
A major addition of state management allows classes obtain provider-stored data from StateManager class without using any stored dependency injections in the instances.

In this use case, we use the BackgroundProvider to provide the data of current background class that is being used, allowing things like FloatingDialog to control the correct background instance whenever it is called to become visible on screen.